+++ Anjeer-Gifts/Models/Categories/Category.cs
@@ -1,0 +1,9 @@
+﻿using Anjeer_Gifts.Models.Commons;
+
+namespace Anjeer_Gifts.Models.Categories;
+
+public class Category : Auditable
+{
+    public string Name { get; set; }
+
+}
