--- Anjeer-Gifts/Models/Category.cs
@@ -1,7 +1,0 @@
-﻿namespace Anjeer_Gifts.Models;
-
-public class Category : Auditable
-{
-    public string Name { get; set; }
-
-}
